+++ src/components/Account/AccountProfilePage.tsx
@@ -1,7 +1,7 @@
 import { promisify } from "util";
 import { IDAOState, IMemberState } from "@daostack/client";
 import * as profileActions from "actions/profilesActions";
-import { enableWeb3ProviderAndWarn, getArc, getWeb3Provider } from "arc";
+import { enableWeb3ProviderAndWarn, getArc, getWeb3Provider, getWeb3ProviderInfo } from "arc";
 
 import BN = require("bn.js");
 import AccountImage from "components/Account/AccountImage";
@@ -126,11 +126,10 @@ class AccountProfilePage extends React.Component<IProps, null> {
         showNotification(NotificationStatus.Failure, "Saving profile failed, please try again");
       }
     } catch (error) {
-      if (web3Provider.isSafe) {
-        showNotification(NotificationStatus.Failure, "We're very sorry, but Gnosis Safe does not support message signing :-(");
-      } else {
-        showNotification(NotificationStatus.Failure, error.message);
-      }
+      // eslint-disable-next-line no-console
+      console.error(error.message);
+      const providerName = getWeb3ProviderInfo(web3Provider).name;
+      showNotification(NotificationStatus.Failure, `We're very sorry, but saving the profile failed.  Your wallet (${providerName}) may not support message signing.`);
     }
     setSubmitting(false);
   }
@@ -152,7 +151,7 @@ class AccountProfilePage extends React.Component<IProps, null> {
           {editing ? (accountProfile && accountProfile.name ? "Edit Profile" : "Set Profile") : "View Profile"}
         </BreadcrumbsItem>
 
-        {dao ? <DaoSidebar {...this.props} dao={dao} /> : ""}
+        {dao ? <DaoSidebar dao={dao} /> : ""}
 
         <div className={css.profileContainer} data-test-id="profile-container">
           { editing && (!accountProfile || !accountProfile.name) ? <div className={css.setupProfile}>In order to evoke a sense of trust and reduce risk of scams, we invite you to create a user profile which will be associated with your current Ethereum address.<br/><br/></div> : ""}