+++ src/components/Proposal/ActionButton.tsx
@@ -4,12 +4,10 @@ import { enableWeb3ProviderAndWarn } from "arc";
 import * as classNames from "classnames";
 import { ActionTypes, default as PreTransactionModal } from "components/Shared/PreTransactionModal";
 import { claimableContributionRewards, hasClaimableRewards } from "lib/util";
-import * as moment from "moment";
 import Tooltip from "rc-tooltip";
 import * as React from "react";
 import { connect } from "react-redux";
 import { IRootState } from "reducers";
-import { closingTime } from "reducers/arcReducer";
 import { proposalEnded } from "reducers/arcReducer";
 import { showNotification } from "reducers/notifications";
 import { IProfileState } from "reducers/profilesReducer";
@@ -24,8 +22,10 @@ interface IExternalProps {
   daoState: IDAOState;
   daoEthBalance: BN;
   detailView?: boolean;
+  expanded?: boolean;
   proposalState: IProposalState;
   rewards: IRewardState;
+  expired: boolean;
 }
 
 interface IStateProps {
@@ -88,12 +88,13 @@ class ActionButton extends React.Component<IProps, IState> {
       daoState,
       daoEthBalance,
       detailView,
+      expired,
+      expanded,
       proposalState,
       rewards,
     } = this.props;
 
     const executable = proposalEnded(proposalState) && !proposalState.executedAt;
-    const expired = closingTime(proposalState).isSameOrBefore(moment());
 
     let beneficiaryHasRewards;
 
@@ -130,6 +131,7 @@ class ActionButton extends React.Component<IProps, IState> {
     const wrapperClass = classNames({
       [css.wrapper]: true,
       [css.detailView]: detailView,
+      [css.expanded]: expanded,
     });
 
     return (